+++ components/Modals/components/AssigneeAndDueDateInfo.module.css
@@ -58,6 +58,5 @@
   }
   .dueDate {
     justify-content: space-between;
-    gap: 1rem;
   }
 }
