+++ components/Modals/components/AssigneeAndDueDateInfo.tsx
@@ -1,8 +1,9 @@
 import React from "react";
 import styles from "./AssigneeAndDueDateInfo.module.css";
-import ProfileIcon from "@/components/header/members/ProfileIcon";
+
 import formatDate, { DateFormat } from "@/utils/formatDate";
 import { TaskInfo } from "../Modal.type";
+import ProfileIcon from "@/components/Members/ProfileIcon";
 
 interface TaskCardInfoProps {
   data: TaskInfo;
